refactor: export/importy global types
