--- apps/color-picker/src/types.d.ts
@@ -1,11 +1,0 @@
-type Color = {
-  id: string;
-  name: string;
-  value: string;
-};
-
-type Theme = {
-  id: string;
-  name: string;
-  colors: Color[];
-};
